+++ pokedex/src/Components/PokeCard/Pokemons.js
@@ -1,0 +1,24 @@
+import React from "react";
+import PokeCard from "./PokeCard";
+
+
+const Pokemons = ({pokemon}) => {
+  
+  return (
+    <div>
+        {
+            pokemon.map((value, index) => {
+                
+                return (
+                    <div key={index}>
+                        <PokeCard thisPokemon={value}/>
+                    </div>
+                )
+            })
+        }
+    </div>
+)
+  
+  
+};
+export default Pokemons;
